[sprint-1] fix int32 overflow: ms-epoch timestamps need BigInteger
SQLite tests passed but Postgres rejected signup — Integer (int4) cannot
hold int(time.time()*1000). All six timestamp columns now BigInteger.
Verified with live auth smoke test against Postgres 16.

+++ analytics-brain/app/models/db_models.py
@@ -2,7 +2,7 @@
 Database models — persistent source of truth.
 Redis is the fast layer. Postgres is what survives a restart.
 """
-from sqlalchemy import String, Float, Integer, Boolean, JSON, Text, ForeignKey
+from sqlalchemy import String, Float, Integer, BigInteger, Boolean, JSON, Text, ForeignKey
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 from app.core.database import Base
 import time
@@ -21,7 +21,7 @@ class MerchantDB(Base):
     logo_url: Mapped[str | None] = mapped_column(String)
     onboarding_status: Mapped[str] = mapped_column(String, default="store_info")
     is_live: Mapped[bool] = mapped_column(Boolean, default=False)
-    created_at: Mapped[int] = mapped_column(Integer, default=lambda: int(time.time() * 1000))
+    created_at: Mapped[int] = mapped_column(BigInteger, default=lambda: int(time.time() * 1000))
 
     products: Mapped[list["ProductDB"]] = relationship(back_populates="merchant")
     orders: Mapped[list["OrderDB"]] = relationship(back_populates="merchant")
@@ -41,7 +41,7 @@ class ProductDB(Base):
     image_urls: Mapped[list] = mapped_column(JSON, default=list)
     is_active: Mapped[bool] = mapped_column(Boolean, default=True)
     qwen_generated_description: Mapped[bool] = mapped_column(Boolean, default=False)
-    created_at: Mapped[int] = mapped_column(Integer, default=lambda: int(time.time() * 1000))
+    created_at: Mapped[int] = mapped_column(BigInteger, default=lambda: int(time.time() * 1000))
 
     merchant: Mapped["MerchantDB"] = relationship(back_populates="products")
 
@@ -55,8 +55,8 @@ class BrandProfileDB(Base):
     logo_analysis: Mapped[dict] = mapped_column(JSON, nullable=False)
     generated_brand: Mapped[dict] = mapped_column(JSON, nullable=False)
     # brand_guard_rules live inside generated_brand JSON
-    created_at: Mapped[int] = mapped_column(Integer, default=lambda: int(time.time() * 1000))
-    updated_at: Mapped[int] = mapped_column(Integer, default=lambda: int(time.time() * 1000))
+    created_at: Mapped[int] = mapped_column(BigInteger, default=lambda: int(time.time() * 1000))
+    updated_at: Mapped[int] = mapped_column(BigInteger, default=lambda: int(time.time() * 1000))
 
 
 class OrderDB(Base):
@@ -69,7 +69,7 @@ class OrderDB(Base):
     total: Mapped[float] = mapped_column(Float, nullable=False)
     status: Mapped[str] = mapped_column(String, default="pending")
     promo_applied: Mapped[str | None] = mapped_column(String)
-    created_at: Mapped[int] = mapped_column(Integer, default=lambda: int(time.time() * 1000))
+    created_at: Mapped[int] = mapped_column(BigInteger, default=lambda: int(time.time() * 1000))
 
     merchant: Mapped["MerchantDB"] = relationship(back_populates="orders")
 
@@ -86,4 +86,4 @@ class DeltaLogDB(Base):
     action_id: Mapped[str] = mapped_column(String, nullable=False)
     patches: Mapped[list] = mapped_column(JSON, nullable=False)
     executed_by: Mapped[str] = mapped_column(String, nullable=False)
-    executed_at: Mapped[int] = mapped_column(Integer, nullable=False, index=True)
+    executed_at: Mapped[int] = mapped_column(BigInteger, nullable=False, index=True)
